+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/meta/snmp/__init__.py
@@ -7,7 +7,7 @@ from ...console import CONTEXT_SETTINGS
 from .generate_profile import generate_profile_from_mibs
 from .translate_profile import translate_profile
 from .validate_mib_filenames import validate_mib_filenames
-from .validators.snmp_validation_click import click_options
+from .validate_snmp_profiles import click_options
 
 ALL_COMMANDS = [generate_profile_from_mibs, translate_profile, validate_mib_filenames, click_options]
