+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/meta/snmp/validators/__init__.py
@@ -1,0 +1,1 @@
+from .schema_validation import validate_profile
